moving middleware logic out of store and into its own file
