+++ lib/src/main/java/com/github/ouchadam/kdux/Store.kt
@@ -1,9 +1,7 @@
 package com.github.ouchadam.kdux
 
-import com.github.ouchadam.kdux.common.CompositeKduxDisposable
 import com.github.ouchadam.kdux.common.KduxDisposable
 import com.github.ouchadam.kdux.common.Middleware
-import com.github.ouchadam.kdux.common.plusAssign
 
 typealias Observer<State> = (State) -> Unit
 typealias Reducer<State, Action> = (Action, currentState: State?) -> State
@@ -11,33 +9,18 @@ typealias Reducer<State, Action> = (Action, currentState: State?) -> State
 typealias ReadState<State> = () -> State
 typealias Dispatch<Action> = (Action) -> Unit
 
-val SYNC_MIDDLEWARE: KduxDisposable = {}
-
-private fun <State, Input> emptyMiddleware(): Middleware<State, Input, Input> {
-    return { input, _ ->
-        { dispatch ->
-            dispatch(input)
-            SYNC_MIDDLEWARE
-        }
-    }
-}
-
 class Store<State, Action, TAction> private constructor(
     private val reducer: Reducer<State, TAction>,
     private val middleware: Middleware<State, Action, TAction>
 ) {
 
     companion object {
-        fun <State, Action> create(reducer: Reducer<State, Action>): Store<State, Action, Action> {
-            return Store(reducer, emptyMiddleware())
-        }
+        fun <State, Action> create(reducer: Reducer<State, Action>) = Store(reducer, emptyMiddleware())
 
         fun <State, InputAction, Action> create(
             reducer: Reducer<State, Action>,
             middleware: Middleware<State, InputAction, Action>
-        ): Store<State, InputAction, Action> {
-            return Store(reducer, middleware)
-        }
+        ) = Store(reducer, middleware)
     }
 
     private val observers = mutableListOf<Observer<State>>()
@@ -77,38 +60,5 @@ class Store<State, Action, TAction> private constructor(
     }
 }
 
-fun <State, Input, Output> syncMiddleware(provider: (Input, ReadState<State?>) -> Output): Middleware<State, Input, Output> {
-    return { input, readState ->
-        { dispatch ->
-            dispatch(provider(input, readState))
-            SYNC_MIDDLEWARE
-        }
-    }
-}
-
-fun <S, M1I, M1O, M2O> combineMiddleware(
-    m1: Middleware<S, M1I, M1O>,
-    m2: Middleware<S, M1O, M2O>
-): Middleware<S, M1I, M2O> = { input, readState ->
-    { dispatch ->
-        val disposables =
-            CompositeKduxDisposable()
-        disposables += m1(input, readState).invoke { result ->
-            disposables += m2(result, readState).invoke { result ->
-                dispatch(result)
-            }
-        }
-        disposables
-    }
-}
-
 class StateHolder<State>(var state: State? = null)
 
-fun<State, Input> loggerMiddleware(log: (String) -> Unit) = { input: Input, readState: ReadState<State?> ->
-    { dispatch: Dispatch<Input> ->
-        log("current state ${readState()}")
-        dispatch(input)
-        log("next state ${readState()}")
-        SYNC_MIDDLEWARE
-    }
-}
